おゆずりのコンポーネントを分割して小さくした
